Added code to keep old msgs for the server, should fix bug where the GUI doesn't always show a accurate representation of the network
also addid an update gui button to force a check for connections and also a redo connections which will connect unconnected modules if possible, HOPEFULLY these buttons wont be used because of the fix before.

Also had reports of modules not reconnecting when shutdown/restarted, working theory is that these modules do not have a shutdown sequence, forcing the OS to terminate the process in a ugly way. Thus yarp server 3 no longer has an accurate network state, can either force a total recheck of ports to solve this which is very time heavy and can cause crashes if people try to use other functions at the same time, or when a module stops existing to simply check the ports associated with that module. Hopefully both wont be needid.

git-svn-id: http://svn.lirec.eu@943 dcc22606-d844-4505-a083-a3dbd7937ca2