Major change in the way that perceptions are received by the AgentCore, in order to improve synchronization.
I've created explicit perception structures, which are added to a set of perception lists by the remote agent. These lists are always processed at the beginning of the core cycle. Before, a property changed perception for instance could occur during any part of the main cycle.

An additional explicit lock mechanism ensures that the core does not process its perceptions while there are messages being read from the remote agent. This implies that perceptions are processed in batches: If two perceptions are received together, the agent will never process one without the other. This solves a problem we had with action canceling, when one of the effects (phone stopped ringing) was perceived and processed before the remaining effects of the action (answering the phone).

git-svn-id: http://svn.lirec.eu@1370 dcc22606-d844-4505-a083-a3dbd7937ca2